requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
beautifulsoup4==4.12.2
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.0.2
networkx==3.2.1
gunicorn==21.2.0
google-genai
python-dotenv
